fix(sms): widen batch total_cost to long/BIGINT to prevent overflow
int 로 계산하면 unit_cost × 건수가 2,147,483,647전을 넘는 순간 조용히 음수가
기록된다. 캐스트는 왼쪽 피연산자에 붙였다 — (long)(a * b) 는 이미 넘친 값을
넓히는 것이라 아무 것도 고치지 못한다.

도달 조건 (잠재 결함이다, 활성 결함이 아니다):
  max-recipients=500 에서는 단가가 4,294,967전(42,949원)을 넘어야 도달한다.
  ① CSM_SMS_BATCH_MAX_RECIPIENTS 상향 — 대량 발송 방향이라 예정된 변경에 가깝다
  ② 단가 입력 오류 — 단가 화면에 자릿수 검증이 없다. CSM-2/CSM-3 으로 닫힌다

운영 데이터 확인 (마이그레이션 안전 근거):
  - total_cost < 0 인 행: 0건
  - total_cost <> unit_cost × (success + unknown) 인 행: 0건
  - 규모: 95행 / MAX(total_cost)=3000전 / MAX(unit_cost)=3000전 / MAX(total_count)=1
  95개 배치가 전부 수신자 1명이다. 500건 상한에 근접한 적조차 없다.
  → 기존 데이터 정정 절차 불필요.

마이그레이션:
  기존 스키마 보정과 같은 방식으로 INFORMATION_SCHEMA 확인 후 조건부 ALTER.
  MySQL 의 INT → BIGINT 는 확대 방향이라 값 손실이 없고 INPLACE 로 처리된다.
  BIGINT 컬럼은 구 JAR 과도 호환되므로 롤백은 이전 JAR 재배포만으로 끝난다.

max-recipients 상향 시 주의사항을 application.properties 에 문서화했다.
오버플로는 이 커밋으로 해소되고, 실제 병목은 순차 발송 처리 시간이다
((건수-1) × send-delay-ms — 10,000건이면 약 16.6분).

회귀 테스트 2건 추가. 수정을 되돌리면 3건이 실패하는 것을 확인했다.

+++ src/main/java/com/coresolution/csm/serivce/SmsBatchService.java
@@ -175,7 +175,19 @@ public class SmsBatchService {
         }
 
         // total_cost 는 Phase 4 차감 금액의 근거다. UNKNOWN 은 환불 금지 원칙이므로 포함한다.
-        int totalCost = unitCost * (success + unknown);
+        //
+        // long 인 이유: int 로 계산하면 unit_cost × 건수가 2,147,483,647전을 넘는 순간
+        // 조용히 음수가 된다. 현재 max-recipients=500 에서는 단가가 4,294,967전(42,949원)을
+        // 넘어야 도달하므로 정상 운영에서는 닿지 않지만, 두 경로로 실현될 수 있다.
+        //   ① CSM_SMS_BATCH_MAX_RECIPIENTS 를 올리는 경우
+        //      — 대량 발송 방향이라 가정이 아니라 예정된 변경에 가깝다.
+        //        상향 시 주의사항은 application.properties 의 max-recipients 주석 참조.
+        //   ② 단가 입력 오류 (단가 화면에 자릿수 검증이 없다)
+        //      — CSM-2 에서 단가 설정 화면·API 가 제거되고 마스터가 플랫폼으로 넘어가면
+        //        이 경로는 닫힌다. 그때는 CSM-3 의 단가 수신 검증이 방어선이 된다.
+        // 캐스트를 왼쪽 피연산자에 붙여야 한다 — (long)(a * b) 는 이미 넘친 값을 넓히는 것이라
+        // 아무 것도 고치지 못한다.
+        long totalCost = (long) unitCost * (success + unknown);
         try {
             jdbcTemplate.update("""
                     UPDATE csm.sms_batch
